// app/src/api/schema.d.ts
export interface components {
    schemas: {
// ...
        RuntimeDevice: {
            /** Device */
            device: string;
            /** Gpu Available */
            gpu_available: boolean;
            /** Gpu Name */
            gpu_name?: string | null;
            failure?: components["schemas"]["DeviceFailure"] | null;
            /** Can Change */
            can_change: boolean;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}